adding a test for config rsync

+++ lib/Conf.pm
@@ -47,6 +47,7 @@ sub refresh
   my($dev,$ino,$mode,$nlink,$uid,$gid,$rdev,$size, $atime,$mtime,$ctime,$blksize,$blocks) = stat("/home/pi/conf");
   my($sec, $min, $hour, $mday, $mon, $year) = localtime($mtime);
   my $client = sprintf("%s-%02d-%02d %02d:%02d", $year + 1900, $mon + 1, $mday, $hour, $min);
+  print "client=$client\n";
 
   ## server
   my $cmd = "/usr/bin/curl -s https://www.ixnay.net/cams/" . $self->name() . "/";
@@ -55,16 +56,19 @@ sub refresh
   if($html =~ /<td><a href="conf\/">conf\/<\/a><\/td><td align="right">(\d+\-\d+\-\d+ \d+:\d+)\s+<\/td>/s) {
     $server = $1;
   }
+  print "server=$server\n";
 
   #
   # rsync?
   #
 
-  if($server ne $client) {
-    my $cmd = "/usr/bin/rsync --timeout=10 -avz --delete uaws2:www/vhosts/ixnay/htdocs/cams/" . $self->{'name'} . "/conf /home/pi";
-    #print "$cmd\n";
-    my $out = `$cmd`;
-    #print "$out\n\n";
+  if($server and $client) {
+    if($server ne $client) {
+      my $cmd = "/usr/bin/rsync --timeout=10 -avz --delete uaws2:www/vhosts/ixnay/htdocs/cams/" . $self->{'name'} . "/conf /home/pi";
+      print "$cmd\n";
+      my $out = `$cmd`;
+      print "$out\n\n";
+    }
   }
 }
